refactor: move ckeditor functionality
